pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "cyed maroof",email = "cyedmaroof7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-openai (>=0.3.0,<0.4.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "openai (>=1.59.7,<2.0.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "langchain-chroma (>=0.2.0,<0.3.0)",
    "pypdf (>=5.1.0,<6.0.0)",
    "langgraph (>=0.2.66,<0.3.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "docx2txt (>=0.8,<0.9)",
    "unstructured[md] (>=0.17.2,<0.18.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "selenium (>=4.29.0,<5.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "google-genai (>=1.7.0,<2.0.0)",
    "llama-index-llms-openai (>=0.3.38,<0.4.0)",
    "llama-index (>=0.12.32,<0.13.0)",
    "llama-index-embeddings-langchain (>=0.3.0,<0.4.0)",
    "llama-index-llms-azure-openai (>=0.3.2,<0.4.0)",
    "pyvis (>=0.3.2,<0.4.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "llama-index-graph-stores-neo4j (>=0.4.6,<0.5.0)",
    "llama-index-llms-langchain (>=0.6.1,<0.7.0)",
    "lightrag-hku (>=1.3.5,<2.0.0)",
    "flagembedding (>=1.3.4,<2.0.0)",
    "graphrag (>=2.2.0,<3.0.0)",
    "ipywidgets (>=8.1.6,<9.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "python-igraph (>=0.11.8,<0.12.0)",
    "leidenalg (>=0.10.2,<0.11.0)",
    "llama-index-extractors-entity (>=0.3.1,<0.4.0)",
    "torch (>=2.7.0,<3.0.0)",
    "llama-index-embeddings-huggingface (>=0.5.4,<0.6.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "llama-index-vector-stores-faiss (>=0.4.0,<0.5.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "gradio (>=5.32.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
